scratch-commit

+++ src/main/java/com/malina_ink/resaleplatform/service/impl/AdServiceImpl.java
@@ -3,6 +3,7 @@ package com.malina_ink.resaleplatform.service.impl;
 import com.malina_ink.resaleplatform.dto.*;
 import com.malina_ink.resaleplatform.entity.Ad;
 import com.malina_ink.resaleplatform.entity.User;
+import com.malina_ink.resaleplatform.exception.AccessErrorException;
 import com.malina_ink.resaleplatform.exception.NotCorrectCostException;
 import com.malina_ink.resaleplatform.exception.ObjectAbsenceException;
 import com.malina_ink.resaleplatform.mapper.AdMapper;
@@ -20,6 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.Objects;
 
 /**
  * Сервис AdsServiceImpl
@@ -73,11 +75,11 @@ public class AdServiceImpl implements AdService {
         adsEntity.setPrice(createAds.getPrice());
         adsEntity.setDescription(createAds.getDescription());
         User author = userRepository.getUserByEmailIgnoreCase(authentication.getName()).orElseThrow(RuntimeException::new); //сделать свое исключение
-        adsEntity.setUser(author);
+        adsEntity.setAuthor(author);
         adRepository.save(adsEntity);
 
         try {
-            String filePath = imageService.uploadAdImage(image, adsEntity.getPk());
+            String filePath = imageService.uploadAdImage(image, adsEntity.getId());
             adsEntity.setAdImage(filePath);
         } catch (IOException e) {
             throw new RuntimeException("Ошибка при загрузке фото");
@@ -105,8 +107,13 @@ public class AdServiceImpl implements AdService {
      * @param adsId идентификатор объявления, не может быть null
      */
     @Override
-    public void deleteAds(Integer adsId) {
+    public void deleteAds(Integer adsId, Authentication authentication) {
         log.info("Вызван метод удаления объявления по идентификатору (id)");
+        //проверка на авторство редактируемого/удаляемого
+        Ad updateAd = adRepository.findById(adsId).orElseThrow(RuntimeException::new);
+        if (!Objects.equals(authentication.getName(), updateAd.getAuthor())) {
+            throw new AccessErrorException("У вас нет прав на данную операцию");
+        }
         adRepository.deleteById(adsId);
     }
 
@@ -118,7 +125,7 @@ public class AdServiceImpl implements AdService {
      * @return возвращает обновленное объявление по идентификатору (id)
      */
     @Override
-    public AdDto updateAds(CreateOrUpdateAdDto createAds, Integer adsId) {
+    public AdDto updateAds(CreateOrUpdateAdDto createAds, Integer adsId, Authentication authentication) {
         if (adsId == null) {
             throw new ObjectAbsenceException("Такого объявления не существует!");
         }
@@ -127,7 +134,12 @@ public class AdServiceImpl implements AdService {
             throw new NotCorrectCostException("Цена должна быть больше 0!");
         }
 
+        //проверка на авторство редактируемого
         Ad updateAd = adRepository.findById(adsId).orElseThrow(RuntimeException::new);
+        if (!Objects.equals(authentication.getName(), updateAd.getAuthor())) {
+            throw new AccessErrorException("У вас нет прав на данную операцию");
+        }
+
         updateAd.setTitle(createAds.getTitle());
         updateAd.setPrice(createAds.getPrice());
         updateAd.setDescription(createAds.getDescription());
@@ -149,7 +161,6 @@ public class AdServiceImpl implements AdService {
         return adMapper.toAdsDto(
                 adRepository.findAllByAuthorIdOrderByIdDesc(userService.getUser(id, authentication).getId())
         );
-
     }
 
     /**
